refactor: url 이미지 모달 컴포넌트 분리
