+++ client/src/components/URLShareModal/URLShareModal.tsx
@@ -1,19 +1,7 @@
-import Button from '../@atom/Button/Button';
-import Textarea from '../@atom/Textarea/Textarea';
-import {
-  DisplayButtonContainer,
-  StyledModal,
-  DisplayButton,
-  SourceCodeContainer,
-  StyledSubTitle,
-  CloseButton,
-  PaletteContainer,
-  ColorSelectButton,
-} from './URLShareModal.styles';
-import DummyButton from '../../assets/images/dummy/button.png';
 import { UserInfo } from '../../types';
-import { MouseEvent, useState } from 'react';
+import { useState } from 'react';
 import URLShareSelectModal from '../URLShareSelectModal/URLShareSelectModal';
+import URLBannerShareModal from '../URLBannerShareModal/URLBannerShareModal';
 
 export interface URLShareModalProps {
   userInfo: UserInfo;
@@ -30,20 +18,6 @@ enum ContentType {
 
 const URLShareModal = ({ userInfo, onClose }: URLShareModalProps) => {
   const [content, setContent] = useState<keyof typeof ContentType>('CHOICE');
-  const sourceCode = `<a href="${window.location.origin}/creator/${userInfo.pageName}" target="_blank"><img height="36" style="border:0px;height:36px;" src="${DummyButton}" border="0" alt="thank you for button" /></a>`;
-
-  const copySourceCode = () => {
-    navigator.clipboard.writeText(sourceCode);
-    alert('소스코드가 복사되었습니다.');
-  };
-
-  const selectAll = ({ currentTarget }: MouseEvent<HTMLTextAreaElement>) => {
-    currentTarget.select();
-  };
-
-  const changeButtonColor = (color: string) => {
-    alert(color);
-  };
 
   const openBannerShareModal = () => {
     setContent('BANNER');
@@ -58,24 +32,7 @@ const URLShareModal = ({ userInfo, onClose }: URLShareModalProps) => {
           openBannerShareModal={openBannerShareModal}
         />
       ) : (
-        <StyledModal onClose={onClose}>
-          <DisplayButtonContainer>
-            <DisplayButton src={DummyButton} alt="" />
-            <PaletteContainer>
-              <ColorSelectButton color="red" onClick={() => changeButtonColor('red')} />
-              <ColorSelectButton color="blue" onClick={() => changeButtonColor('blue')} />
-              <ColorSelectButton color="purple" onClick={() => changeButtonColor('purple')} />
-              <ColorSelectButton color="green" onClick={() => changeButtonColor('green')} />
-              <ColorSelectButton color="yellow" onClick={() => changeButtonColor('yellow')} />
-            </PaletteContainer>
-          </DisplayButtonContainer>
-          <SourceCodeContainer>
-            <StyledSubTitle>복사 붙여넣기 코드</StyledSubTitle>
-            <Textarea value={sourceCode} readOnly onClick={selectAll} />
-          </SourceCodeContainer>
-          <Button onClick={copySourceCode}>소스코드 복사하기</Button>
-          <CloseButton onClick={onClose}>창 닫기</CloseButton>
-        </StyledModal>
+        <URLBannerShareModal userInfo={userInfo} onClose={onClose} />
       )}
     </>
   );
